+++ src/lib/waku_store/history_rpc.ts
@@ -2,31 +2,42 @@ import { Reader } from 'protobufjs/minimal';
 import { v4 as uuid } from 'uuid';
 
 import * as proto from '../../proto/waku/v2/store';
-import { DefaultContentTopic } from '../waku_message';
-import { DefaultPubsubTopic } from '../waku_relay';
+
+export enum Direction {
+  BACKWARD = 'backward',
+  FORWARD = 'forward',
+}
+
+export interface Options {
+  contentTopics: string[];
+  cursor?: proto.Index;
+  pubsubTopic: string;
+  direction: Direction;
+  pageSize: number;
+}
 
 export class HistoryRPC {
   public constructor(public proto: proto.HistoryRPC) {}
 
-  static createQuery(
-    contentTopics: string[] = [DefaultContentTopic],
-    cursor?: proto.Index,
-    pubsubTopic: string = DefaultPubsubTopic
-  ): HistoryRPC {
+  /**
+   * Create History Query.
+   */
+  static createQuery(options: Options): HistoryRPC {
+    const direction = directionToProto(options.direction);
     const pagingInfo = {
-      pageSize: 10,
-      cursor,
-      direction: proto.PagingInfo_Direction.DIRECTION_FORWARD,
+      pageSize: options.pageSize,
+      cursor: options.cursor,
+      direction,
     };
 
-    const contentFilters = contentTopics.map((contentTopic) => {
+    const contentFilters = options.contentTopics.map((contentTopic) => {
       return { contentTopic };
     });
 
     return new HistoryRPC({
       requestId: uuid(),
       query: {
-        pubsubTopic,
+        pubsubTopic: options.pubsubTopic,
         contentFilters,
         pagingInfo,
         startTime: undefined,
@@ -51,5 +62,16 @@ export class HistoryRPC {
 
   get response(): proto.HistoryResponse | undefined {
     return this.proto.response;
+  }
+}
+
+function directionToProto(direction: Direction): proto.PagingInfo_Direction {
+  switch (direction) {
+    case Direction.BACKWARD:
+      return proto.PagingInfo_Direction.DIRECTION_BACKWARD_UNSPECIFIED;
+    case Direction.FORWARD:
+      return proto.PagingInfo_Direction.DIRECTION_FORWARD;
+    default:
+      return proto.PagingInfo_Direction.DIRECTION_BACKWARD_UNSPECIFIED;
   }
 }